Changes to utilize remaining historical figure XML data.
Includes entity reputations, related sites, skills and artifacts held. Current and Used identities are incorrect in the XML so they are hidden from the user.

Event Changes.
Filled in the AddHFHFLink event so it display accurately in the logs hopefully.
Added the agreement made event.

Added auto map resizing.

Some future proofing of the XMLParser so changes to the XML data won't cause crashes.

Added artifacts to the advanced search.

Added simple searching options for vampires and werebeasts.