Fixes to historical figure curse filtering.
Curses (vampires, werebeasts) were being filtered by searching for exact matches to the curse name (E.G. "VAMPIRE"), however this did not match the curse strings in the active interactions. Filtering (and historical figure naming based on curses) was fixed by using a search through all active interactions for strings that contained the sub-string of the curse name.

This method for determining if a historical figure is cursed has the issue that it searches through all active interactions for the respective sub-strings, when really only an initial match is necessary. This performance hit however does not seem significant in comparison the simplicity of the solution.