Fixed a crash related to civilisations lost to the end of stream.
If while parsing a history file, the file ends with the name of a civilisation, end of stream will trigger and the civilisation will never been parsed. This will later result in a crash when attempting to print the `WorldStatsPrinter`. I addressed this by checking if a remaining current line is left unparsed post closing the history file, and then testing and parsing it as a civilisation. This needed a new EoS oriented function to avoid calling `ReadLine` while handling the civ, however due to it being the last line does not require worship or leadership parsing.